First batch of fixes.
